refactor: split cms components into submodules
