+++ packages/ui/src/components/cms/page-builder/ComponentEditor.tsx
@@ -2,18 +2,11 @@
 "use client";
 
 import type { PageComponent } from "@acme/types";
-import { memo, Suspense } from "react";
-import {
-  Button,
-  Input,
-  Accordion,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "../../atoms/shadcn";
-import editorRegistry from "./editorRegistry";
+import { memo } from "react";
+import { Accordion } from "../../atoms/shadcn";
+import LayoutPanel from "./panels/LayoutPanel";
+import ContentPanel from "./panels/ContentPanel";
+import InteractionsPanel from "./panels/InteractionsPanel";
 import useComponentInputs from "./useComponentInputs";
 import useComponentResize from "./useComponentResize";
 
@@ -46,254 +39,38 @@ function ComponentEditor({ component, onChange, onResize }: Props) {
   const { handleInput } = useComponentInputs(onChange);
   const { handleResize, handleFullSize } = useComponentResize(onResize);
 
-  const Specific = editorRegistry[component.type];
-
-  const layout = (
-    <div className="space-y-2">
-      {(["Desktop", "Tablet", "Mobile"] as const).map((vp) => (
-        <div key={vp} className="space-y-2">
-          <div className="flex items-end gap-2">
-            <Input
-              label={`Width (${vp})`}
-              placeholder="e.g. 100px or 50%"
-              title="CSS width value with units"
-              value={(component as any)[`width${vp}`] ?? ""}
-              onChange={(e) => handleResize(`width${vp}`, e.target.value)}
-            />
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => handleFullSize(`width${vp}`)}
-            >
-              Full width
-            </Button>
-          </div>
-          <div className="flex items-end gap-2">
-            <Input
-              label={`Height (${vp})`}
-              placeholder="e.g. 1px or 1rem"
-              title="CSS height value with units"
-              value={(component as any)[`height${vp}`] ?? ""}
-              onChange={(e) => handleResize(`height${vp}`, e.target.value)}
-            />
-            <Button
-              type="button"
-              variant="outline"
-              onClick={() => handleFullSize(`height${vp}`)}
-            >
-              Full height
-            </Button>
-          </div>
-        </div>
-      ))}
-      <Select
-        value={component.position ?? ""}
-        onValueChange={(v) => handleInput("position", v || undefined)}
-      >
-        <SelectTrigger title="CSS position property">
-          <SelectValue placeholder="Position" />
-        </SelectTrigger>
-        <SelectContent>
-          <SelectItem value="relative">relative</SelectItem>
-          <SelectItem value="absolute">absolute</SelectItem>
-        </SelectContent>
-      </Select>
-      {component.position === "absolute" && (
-        <>
-          <Input
-            label="Top"
-            placeholder="e.g. 10px"
-            title="CSS top offset with units"
-            value={component.top ?? ""}
-            onChange={(e) => handleResize("top", e.target.value)}
-          />
-          <Input
-            label="Left"
-            placeholder="e.g. 10px"
-            title="CSS left offset with units"
-            value={component.left ?? ""}
-            onChange={(e) => handleResize("left", e.target.value)}
-          />
-        </>
-      )}
-    </div>
-  );
-
-  const spacing = (
-    <div className="space-y-2">
-      {(["Desktop", "Tablet", "Mobile"] as const).map((vp) => (
-        <div key={vp} className="space-y-2">
-          <Input
-            label={`Margin (${vp})`}
-            placeholder="e.g. 1rem"
-            title="CSS margin value with units"
-            value={(component as any)[`margin${vp}`] ?? ""}
-            onChange={(e) => handleResize(`margin${vp}`, e.target.value)}
-          />
-          <Input
-            label={`Padding (${vp})`}
-            placeholder="e.g. 1rem"
-            title="CSS padding value with units"
-            value={(component as any)[`padding${vp}`] ?? ""}
-            onChange={(e) => handleResize(`padding${vp}`, e.target.value)}
-          />
-        </div>
-      ))}
-      <Input
-        label="Margin"
-        placeholder="e.g. 1rem"
-        title="Global CSS margin value with units"
-        value={component.margin ?? ""}
-        onChange={(e) => handleInput("margin", e.target.value)}
-      />
-      <Input
-        label="Padding"
-        placeholder="e.g. 1rem"
-        title="Global CSS padding value with units"
-        value={component.padding ?? ""}
-        onChange={(e) => handleInput("padding", e.target.value)}
-      />
-      {"gap" in component && (
-        <Input
-          label="Gap"
-          placeholder="e.g. 1rem"
-          title="Gap between items"
-          value={(component as any).gap ?? ""}
-          onChange={(e) => handleInput("gap", e.target.value)}
-        />
-      )}
-    </div>
-  );
-
-  const content = (
-    <div className="space-y-2">
-      {("minItems" in component || "maxItems" in component) && (
-        <>
-          <Input
-            label="Min Items"
-            type="number"
-            title="Minimum number of items"
-            value={(component as any).minItems ?? ""}
-            onChange={(e) => {
-              const val =
-                e.target.value === "" ? undefined : Number(e.target.value);
-              if (val === undefined) {
-                handleInput("minItems", undefined);
-                return;
-              }
-              const max = (component as any).maxItems;
-              const patch: Partial<PageComponent> = { minItems: val };
-              if (max !== undefined && val > max) {
-                patch.maxItems = val;
-              }
-              onChange(patch);
-            }}
-            min={0}
-            max={(component as any).maxItems ?? undefined}
-          />
-          <Input
-            label="Max Items"
-            type="number"
-            title="Maximum number of items"
-            value={(component as any).maxItems ?? ""}
-            onChange={(e) => {
-              const val =
-                e.target.value === "" ? undefined : Number(e.target.value);
-              if (val === undefined) {
-                handleInput("maxItems", undefined);
-                return;
-              }
-              const min = (component as any).minItems;
-              const patch: Partial<PageComponent> = { maxItems: val };
-              if (min !== undefined && val < min) {
-                patch.minItems = val;
-              }
-              onChange(patch);
-            }}
-            min={(component as any).minItems ?? 0}
-          />
-        </>
-      )}
-      {("desktopItems" in component ||
-        "tabletItems" in component ||
-        "mobileItems" in component) && (
-        <>
-          <Input
-            label="Desktop Items"
-            type="number"
-            title="Items shown on desktop"
-            value={(component as any).desktopItems ?? ""}
-            onChange={(e) =>
-              handleInput(
-                "desktopItems",
-                e.target.value === "" ? undefined : Number(e.target.value)
-              )
-            }
-            min={0}
-          />
-          <Input
-            label="Tablet Items"
-            type="number"
-            title="Items shown on tablet"
-            value={(component as any).tabletItems ?? ""}
-            onChange={(e) =>
-              handleInput(
-                "tabletItems",
-                e.target.value === "" ? undefined : Number(e.target.value)
-              )
-            }
-            min={0}
-          />
-          <Input
-            label="Mobile Items"
-            type="number"
-            title="Items shown on mobile"
-            value={(component as any).mobileItems ?? ""}
-            onChange={(e) =>
-              handleInput(
-                "mobileItems",
-                e.target.value === "" ? undefined : Number(e.target.value)
-              )
-            }
-            min={0}
-          />
-        </>
-      )}
-      {"columns" in component && (
-        <Input
-          label="Columns"
-          type="number"
-          title="Number of columns"
-          value={(component as any).columns ?? ""}
-          onChange={(e) =>
-            handleInput(
-              "columns",
-              e.target.value === "" ? undefined : Number(e.target.value)
-            )
-          }
-          min={(component as any).minItems}
-          max={(component as any).maxItems}
-        />
-      )}
-      <Suspense fallback={<p className="text-muted text-sm">Loading...</p>}>
-        {Specific ? (
-          <Specific component={component} onChange={onChange} />
-        ) : (
-          <p className="text-muted text-sm">No editable props</p>
-        )}
-      </Suspense>
-    </div>
-  );
-
   return (
     <Accordion
       items={[
-        { title: "Layout", content: layout },
-        { title: "Spacing", content: spacing },
-        { title: "Content", content: content },
+        {
+          title: "Layout",
+          content: (
+            <LayoutPanel
+              component={component}
+              handleInput={handleInput}
+              handleResize={handleResize}
+              handleFullSize={handleFullSize}
+            />
+          ),
+        },
+        {
+          title: "Content",
+          content: (
+            <ContentPanel
+              component={component}
+              onChange={onChange}
+              handleInput={handleInput}
+            />
+          ),
+        },
+        {
+          title: "Interactions",
+          content: <InteractionsPanel component={component} />,
+        },
       ]}
     />
   );
 }
 
 export default memo(ComponentEditor);
+
